requirements.txt
matplotlib==3.0.3
requests==2.21.0
pandas==0.24.2
lxml==4.1.1
Flask==1.0.2
numpy==1.16.2
python-dotenv==0.10.5
keras==2.3.1
tensorflow==2.0.0
